Move NotifierSelection to VM Reports directory
